# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Hermes Automation API
  version: 1.0.0
  summary: REST API for triggering and monitoring TISS billing automations.
  description: |
    External-system entry point to the Hermes automation pipeline. A caller can:

    - upload one or more TISS files (`.zip` / `.xml`) to create a job;
    - optionally pre-supply a `platformCredentialId` to skip the human approval step;
    - poll job status or stream live events via SSE;
    - manage platform credentials programmatically.

    All endpoints are scoped per API key owner. Resources are isolated per user; an
    API key cannot read or modify resources owned by another user.
  contact:
    name: Webstar
    email: dev@webstar.studio
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://hermes.example.com
    description: Production (replace with your deployment)
...
security:
  - apiKey: []
paths:
  /api/v1/jobs:
    post:
      tags: [Jobs]
      operationId: createJob
      summary: Create a job and enqueue the automation
      description: |
        Accepts one or more files via `multipart/form-data` and starts the durable
        workflow. If `platformCredentialId` is provided, the human-validation step is
        auto-approved (signaled by the `validation_approved` event with
        `autoApproved: true` in its payload).

        Files are saved to the configured object store (R2 in production, local fs in
        development). Per-file limits — `.zip` or `.xml` only, ≤ 25 MB each, 1–50
        files per request.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: array
                  minItems: 1
                  maxItems: 50
                  items:
                    type: string
                    format: binary
                  description: One or more TISS `.zip` or `.xml` files.
                flowType:
                  $ref: "#/components/schemas/JobFlowType"
                platformId:
                  $ref: "#/components/schemas/PlatformId"
                platformCredentialId:
                  type: string
                  description: |
                    Optional. When set, the workflow auto-approves the validation
                    step and proceeds directly to login + submission. Must be a
                    credential owned by the API key's user.
            encoding:
              file:
                contentType: application/zip, application/xml
      responses:
        "200":
          description: Job created and workflow enqueued.
          content:
            application/json:
              schema:
                type: object
                required: [ok, jobId, runId, fileCount, autoApproved]
                properties:
                  ok: { const: true }
                  jobId: { type: string, format: uuid }
                  runId:
                    type: string
                    description: Workflow run id (used internally by the SSE stream).
                  fileCount: { type: integer, minimum: 1 }
                  autoApproved:
                    type: boolean
                    description: True when `platformCredentialId` was provided.
        "400":
          description: Validation error (missing file, too many files, unknown credential, etc.).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    get:
      tags: [Jobs]
      operationId: listJobs
      summary: List jobs owned by the API key holder
      description: Returns up to 50 most recent jobs for the authenticated user.
      responses:
        "200":
          description: Job list (newest first).
          content:
            application/json:
              schema:
                type: object
                required: [ok, jobs]
                properties:
                  ok: { const: true }
                  jobs:
                    type: array
                    items: { $ref: "#/components/schemas/JobListEntry" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /api/v1/jobs/{jobId}/cancel:
    parameters:
      - $ref: "#/components/parameters/JobId"
    post:
      tags: [Jobs]
      operationId: cancelJob
      summary: Cancel an in-flight job
      description: |
        Cancels the underlying durable workflow if the job is still in a live status
        (`uploaded`, `awaiting_validation`, `approved`, `running`). The job row is
        preserved (status set to `failed`, `errorMessage` set to "Cancelado via API")
        so the timeline remains auditable. Idempotent: calling on a terminal job
        returns `200` with `alreadyTerminal: true`.
      responses:
        "200":
          description: Cancellation accepted (or job was already terminal).
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    required: [ok]
                    properties:
                      ok: { const: true }
                  - type: object
                    required: [ok, alreadyTerminal, status]
                    properties:
                      ok: { const: true }
                      alreadyTerminal: { const: true }
                      status: { $ref: "#/components/schemas/JobStatus" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/platform-credentials/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Platform credentials]
      operationId: getPlatformCredential
      summary: Read one credential
      responses:
        "200":
          description: Credential snapshot.
          content:
            application/json:
              schema:
                type: object
                required: [ok, credential]
                properties:
                  ok: { const: true }
                  credential: { $ref: "#/components/schemas/PlatformCredential" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Platform credentials]
      operationId: updatePlatformCredential
      summary: Update label / username / password
      description: |
        Any subset of `label`, `username`, `password` can be sent. Sending `password`
        re-encrypts with the current `CREDENTIAL_ENCRYPTION_KEY`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CredentialPatch" }
      responses:
        "200":
          description: Credential updated.
          content:
            application/json:
              schema:
                type: object
                required: [ok, credential]
                properties:
                  ok: { const: true }
                  credential: { $ref: "#/components/schemas/PlatformCredential" }
        "400":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
          description: Invalid body.
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Duplicate label.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    delete:
      tags: [Platform credentials]
      operationId: deletePlatformCredential
      summary: Delete a credential
      description: |
        Fails with `409 CREDENTIAL_IN_USE` when an existing job still references the
        credential — update or create a new one instead.
      responses:
        "200":
          description: Credential deleted.
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok: { const: true }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Credential in use by an existing job.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: hapi_*
      description: |
        API key issued under **Settings → API Keys** (`/app/settings/api-keys`).

        Format: `hapi_` + 32-character base32 body (≈ 37 chars total). The plaintext
        secret is shown **only once** at creation; the server stores `sha256(secret)`
        and discards the raw value.

        Send the key as `Authorization: Bearer hapi_xxxxxxxxxx…`.
  parameters:
    JobId:
      in: path
      name: jobId
      required: true
      schema: { type: string, format: uuid }
  responses:
    Unauthorized:
      description: Missing, malformed, revoked, or expired API key.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: API key owner is not in `approved` status.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Resource not found, or not owned by the caller.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    JobStatus:
      type: string
      enum: [uploaded, awaiting_validation, approved, running, login_succeeded, failed]
      description: |
        Job lifecycle. Terminal states: `login_succeeded`, `failed`.

        Transitions:
        - `uploaded` → `awaiting_validation` (after TISS extraction)
        - `awaiting_validation` → `approved` (human approval **or** auto-approve)
        - `approved` → `running` (login + submit started)
        - `running` → `login_succeeded` | `failed`
    JobFlowType:
      type: string
      enum: [short, complete]
      default: short
      description: |
        - `short`: upload the `.zip` to the portal and confirm submission.
        - `complete`: open *Digitar Guia* and key in each guide field-by-field.
    PlatformId:
      type: string
      enum: [orizon_fature]
      default: orizon_fature
    Job:
      type: object
      required:
        [
          id, userId, status, flowType, runId, platformId, platformCredentialId,
          validationHookToken, errorMessage, createdAt, updatedAt,
        ]
      properties:
        id: { type: string, format: uuid }
        userId: { type: string }
        status: { $ref: "#/components/schemas/JobStatus" }
        flowType: { $ref: "#/components/schemas/JobFlowType" }
        runId:
          type: [string, "null"]
          description: Workflow run id (set right after enqueue).
        platformId:
          oneOf:
            - { $ref: "#/components/schemas/PlatformId" }
            - type: "null"
        platformCredentialId:
          type: [string, "null"]
        validationHookToken:
          type: [string, "null"]
          description: Internal — used by the resume-validation flow.
        errorMessage:
          type: [string, "null"]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    JobListEntry:
      allOf:
        - $ref: "#/components/schemas/Job"
        - type: object
          required: [file, fileCount, tiss]
          properties:
            file:
              oneOf:
                - type: object
                  properties:
                    fileName: { type: string }
                  required: [fileName]
                - type: "null"
            fileCount: { type: integer, minimum: 0 }
            tiss:
              oneOf:
                - { $ref: "#/components/schemas/TissDocument" }
                - type: "null"
...
    TissDocument:
      type: object
      required:
        [
          id, jobId, standardVersion, transactionType, providerName,
          providerRegister, operatorRegister, batchNumber, guideCount,
          totalAmount, beneficiaryNames, rawSummary, validatedData,
          createdAt, updatedAt,
        ]
      properties:
        id: { type: string, format: uuid }
        jobId: { type: string, format: uuid }
        standardVersion:
          type: [string, "null"]
          examples: ["4.02.00"]
        transactionType:
          type: [string, "null"]
          examples: ["ENVIO_LOTE_GUIAS"]
        providerName: { type: [string, "null"] }
        providerRegister: { type: [string, "null"] }
        operatorRegister:
          type: [string, "null"]
          description: ANS operator registration (typically 6 digits).
        batchNumber: { type: [string, "null"] }
        guideCount:
          type: [string, "null"]
          description: Number of guides in the batch (stored as text).
        totalAmount:
          type: [string, "null"]
          description: Total invoice amount in BRL, two decimals.
        beneficiaryNames:
          type: array
          items: { type: string }
          description: Up to 50 beneficiary names extracted from the XML (PII — handle accordingly).
        rawSummary:
          type: object
          description: Structured TISS summary (versioned shape — see `expanded` field).
          additionalProperties: true
        validatedData:
          oneOf:
            - type: object
              additionalProperties: true
            - type: "null"
          description: Free-form payload supplied at validation time.
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    PlatformCredential:
      type: object
      required: [id, platformId, label, usernameMasked]
      properties:
        id: { type: string, format: uuid }
        platformId: { $ref: "#/components/schemas/PlatformId" }
        label: { type: string, minLength: 2, maxLength: 80 }
        usernameMasked:
          type: string
          description: Username with the middle masked (e.g. `us***om`). The plaintext is never returned.
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    CredentialPatch:
      type: object
      minProperties: 1
      properties:
        label: { type: string, minLength: 2, maxLength: 80 }
        username: { type: string, minLength: 1, maxLength: 160 }
        password:
          type: string
          minLength: 1
          maxLength: 512
          format: password
    ErrorResponse:
      type: object
      required: [ok, error]
      properties:
        ok: { const: false }
        error:
          type: object
          required: [message]
          properties:
            code:
              type: string
              description: |
                Stable, machine-readable code. Known values include:
                `UNAUTHORIZED`, `PENDING_APPROVAL`, `NOT_FOUND`, `INVALID_BODY`,
                `FILE_REQUIRED`, `TOO_MANY_FILES`, `UPLOAD_FAILED`,
                `CREDENTIAL_NOT_FOUND`, `DUPLICATE_LABEL`, `CREDENTIAL_IN_USE`,
                `RUN_NOT_READY`, `RUN_NOT_FOUND`.
            message: { type: string }